Little cleanup in intersection unit tests.
